refactor(server): 拆 PA 巨石为 primary/ 子包 + service 分层 + routes 去业务化
- PA 巨石 primary_agent_service.py 1671→666 行（−60%）：按职责簇抽到 primary/ 子包
  （helpers/lifecycle/rendering/rotation/watcher/dispatcher/delivery），门面持状态、
  子模块接 svc 自由函数，re-export 保测试符号，行为冻结。
- services/base.py 正名 subprocess_utils.py（名实相符）；新增 BaseService ABC 契约。
- routes/settings.py 去业务化：auth/OS 操作下沉 MainConfigService/SystemService，
  route 不再起子进程/写文件。
- state/manager.py 拆 loader/broadcaster，退化为门面。
- 修复 guide_service._parse_frontmatter_manual 真实 bug（标准 frontmatter 被错解析）。
